requirements.txt
sense-hat>=2.2.0
RPi.GPIO>=0.7.0 
flask >=1.1.2
psutil >= 5.7.0
google-cloud-logging>=1.15.0
